Made the callback for sub categories + started styling. Moved product controller code into own file
